Move `_calculate_sliding_sync_table_changes(...)` after we assign `stream_ordering` to events
See https://github.com/element-hq/synapse/pull/17512#discussion_r1725728637
